refactor(foundation): move intelligent_behavior base types (A2)
Phase 6 / Sub-phase A2: Move pure-data types from intelligent_behavior to foundation.
These types are used by non-deprecated core code (openapi, reality, voice,
failure_analysis, graph modules) but have zero crate-internal dependencies.

Moved to mockforge-foundation::intelligent_behavior:
- Persona (w/ get_numeric_trait, get_trait methods)
- LlmGenerationRequest (w/ builder methods)
- Request (MockAI HTTP request shape)
- Response (MockAI HTTP response shape)

Core modules now re-export these from foundation:
- config.rs: Persona
- types.rs: LlmGenerationRequest
- mockai.rs: Request, Response

Features propagated:
- mockforge-foundation now has a `schema` feature that enables
  schemars::JsonSchema derives; forwarded from mockforge-core/schema
